token refreshing 5

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required,get_jwt
 from models.item import ItemModel
 
 
@@ -38,7 +38,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required(optional=False)
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message':'admin previlage required'},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
